backend/requirements.txt
numpy>=1.26
pydantic>=2.5
pydantic-settings>=2.0
aiokafka>=0.10.0
pytest>=7.4
pytest-asyncio>=0.23
sqlalchemy>=2.0
alembic>=1.13
python-dotenv>=1.0
psycopg[binary]>=3.0.7
asyncpg>=0.29
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
